Begin extracting functionality to a library
Extract the creation and removal of entities to an external library.
This will help keep the executable's code clean as the program grows and
also helps reduce duplicate code in the executable. Also, a shared
library will allow other people to interact with the mend database from
their own applications.